there is no need to have an abstract class with a single implementation
